correcting int/double type

+++ statstracker.h
@@ -203,8 +203,8 @@ public:
     vector< vector<int> > getTimeSpentGatheringWithDevice() { return timeSpentGatheringWithDevice; };
     vector< vector<vector< vector<double> > > > getTimeSpentGatheringWithDeviceByRes() { return timeSpentGatheringWithDeviceByAgentByRes; };
     vector< vector<vector<int> > > getTimeSpentGatheringWithDeviceByGroup() { return timeSpentGatheringWithDeviceByGroup; };
-    vector< vector< vector< double > > > getTimeSpentGatheringWithoutDevice() { return timeSpentGatheringWithoutDevice; };
-    vector< vector<vector<int> > > getTimeSpentGatheringWithoutDeviceByRes() { return timeSpentGatheringWithoutDeviceByAgentByRes; };
+    vector< vector< vector< int > > > getTimeSpentGatheringWithoutDevice() { return timeSpentGatheringWithoutDevice; };
+    vector< vector<vector<double> > > getTimeSpentGatheringWithoutDeviceByRes() { return timeSpentGatheringWithoutDeviceByAgentByRes; };
     vector<vector<int> > getTimeSpentGatheringWithoutDeviceByGroup() { return timeSpentGatheringWithoutDeviceByGroup; };
     vector< vector<int> > getTimeSpentMakingDevices() { return timeSpentMakingDevices; };
     vector<vector<vector<int> > > getTimeSpentMakingDevicesByGroup() { return timeSpentMakingDevicesByGroup; };
